requirements.txt
asgiref==3.5.0
boto3==1.21.27
botocore==1.24.27
Django==4.0.3
django-s3direct==2.0.0
django-storages==1.12.3
djangorestframework==3.13.1
jmespath==1.0.0
Pillow==9.0.1
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
